Moved get_schema_hash to repository
