+++ tests/test_models.py
@@ -222,13 +222,6 @@ class EntityTests:
         for field in data:
             assert field in representation
 
-    @patch('eventsourcing_helpers.models.Entity._get_model_representation')
-    def test_hash_gets_model_representation(self, get_model_representation_mock):  # noqa
-        get_model_representation_mock.return_value = 'a'
-        self.aggregate_root.get_schema_hash()
-
-        get_model_representation_mock.assert_called_once_with()
-
 
 class EntityDictTests:
